refactoring + openapi endpoint implementation
